+++ libs/react-components/src/community/components/layout/header/header/header.tsx
@@ -1,11 +1,11 @@
 import cn from 'classnames';
 import React from 'react';
 
+import { Affix } from '../../../../../tedi/components/affix/affix';
 import { Col, Row } from '../../../../../tedi/components/grid';
 import Print from '../../../../../tedi/components/print/print';
 import { Layouts, useLayout } from '../../../../helpers';
 import { IntentionalAny } from '../../../../types';
-import Affix from '../../../affix/affix';
 import Anchor from '../../../anchor/anchor';
 import { LayoutContext } from '../../layout-context';
 import { useSidenavRendered } from '../../sidenav/utility';