+++ libs/react-components/src/community/components/table-of-contents/table-of-contents.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 
+import { Affix } from '../../../tedi/components/affix/affix';
 import ScrollVisibility from '../../../tedi/components/scroll-visibility/scroll-visibility';
 import { StretchContent } from '../../../tedi/components/stretch-content/stretch-content';
 import { useLabels } from '../../../tedi/providers/label-provider';
 import { Layouts, useLayout } from '../../helpers';
-import Affix from '../affix/affix';
 import { Card, CardContent } from '../card';
 import { ModalProps, ModalProviderProps } from '../modal';
 import { TableOfContentsItemProps } from './table-of-contents-item';