+++ libs/react-components/src/tedi/components/form/textarea/textarea.stories.tsx
@@ -1,4 +1,5 @@
 import { Meta, StoryFn, StoryObj } from '@storybook/react';
+import { useState } from 'react';
 
 import { Col, Row } from '../../grid';
 import { Text } from '../../typography/text/text';
@@ -68,6 +69,12 @@ const TemplateSizes: StoryFn<TextAreaProps> = (args) => {
       ))}
     </div>
   );
+};
+
+const TemplateTextValue: StoryFn<TextAreaProps> = (args) => {
+  const { value, ...props } = args;
+  const [text, setText] = useState(value ?? '');
+  return <TextArea value={text} onChange={(t) => setText(t)} {...props} />;
 };
 
 export const Default: Story = {
@@ -142,6 +149,7 @@ export const Error: Story = {
 };
 
 export const TextValue: Story = {
+  render: TemplateTextValue,
   args: {
     id: 'example-1',
     label: 'Label',
@@ -154,5 +162,13 @@ export const Placeholder: Story = {
     id: 'example-1',
     label: 'Label',
     placeholder: 'Text value',
+  },
+};
+
+export const DefaultValue: Story = {
+  args: {
+    id: 'example-1',
+    label: 'Label',
+    defaultValue: 'Text value',
   },
 };