+++ libs/react-components/src/tedi/components/separator/separator.stories.tsx
@@ -148,6 +148,19 @@ export const VerticalDottedSmall: Story = {
   args: { axis: 'vertical', variant: 'dotted-small', color: 'accent', isStretched: true },
 };
 
+export const HorizontalDottedSeparator: Story = {
+  render: () => (
+    <Row>
+      <Col lg={3} md={6} sm={12}>
+        <VerticalSpacing size={2}>
+          <Separator axis="horizontal" variant="dotted" color="accent" />
+          <Separator axis="horizontal" variant="dotted-small" color="accent" />
+        </VerticalSpacing>
+      </Col>
+    </Row>
+  ),
+};
+
 export const DotOnly: Story = {
   render: DotOnlyTemplate,
   args: { spacing: 0.5 },
@@ -172,7 +185,7 @@ const InlineSeparatorTemplate: StoryFn<SeparatorProps> = (args) => (
     </Text>
     <Text>
       Lorem ipsum dolor sit, amet
-      <Separator color="secondary" spacing={0.5} variant="dot-only" dotSize="small" {...args} />
+      <Separator {...args} color="secondary" spacing={0.5} variant="dot-only" dotSize="small" />
       consectetur adipisicing elit.
     </Text>
   </VerticalSpacing>