+++ libs/react-components/src/tedi/components/truncate/truncate.stories.tsx
@@ -10,7 +10,7 @@ import { Truncate } from './truncate';
 
 const meta: Meta<typeof Truncate> = {
   component: Truncate,
-  title: 'Tedi-ready/Components/Helpers/Truncate',
+  title: 'Tedi-ready/Content/Truncate',
   parameters: {
     status: {
       type: [{ name: 'breakpointSupport', url: '?path=/docs/helpers-usebreakpointprops--usebreakpointprops' }],