+++ libs/react-components/src/tedi/components/typography/heading/heading.stories.tsx
@@ -71,3 +71,62 @@ export const CustomModifier: Story = {
     </>
   ),
 };
+
+export const HeadingWithoutModifiers: Story = {
+  render: () => <Heading>Heading</Heading>,
+};
+
+export const SemanticHeadings: Story = {
+  render: () => (
+    <>
+      <Row>
+        <Col>
+          <h1>Heading 1</h1>
+        </Col>
+        <Col>
+          <Heading element="h1">Heading 1</Heading>
+        </Col>
+      </Row>
+      <Row>
+        <Col>
+          <h2>Heading 2</h2>
+        </Col>
+        <Col>
+          <Heading element="h2">Heading 2</Heading>
+        </Col>
+      </Row>
+      <Row>
+        <Col>
+          <h3>Heading 3</h3>
+        </Col>
+        <Col>
+          <Heading element="h3">Heading 3</Heading>
+        </Col>
+      </Row>
+      <Row>
+        <Col>
+          <h4>Heading 4</h4>
+        </Col>
+        <Col>
+          <Heading element="h4">Heading 4</Heading>
+        </Col>
+      </Row>
+      <Row>
+        <Col>
+          <h5>Heading 5</h5>
+        </Col>
+        <Col>
+          <Heading element="h5">Heading 5</Heading>
+        </Col>
+      </Row>
+      <Row>
+        <Col>
+          <h6>Heading 6</h6>
+        </Col>
+        <Col>
+          <Heading element="h6">Heading 6</Heading>
+        </Col>
+      </Row>
+    </>
+  ),
+};